requirements.txt
PyQt5
fpdf
Pillow
numpy
notebook
jupyterlab
voila